test: Use PRId64 for portability
POSIX defines PRId64/PRIu64/PRIx64 for printing 64-bit values in a
portable way. Replace a few references to %ld to remove the assumption
about the size of a long. Similarly, use %z with size_t arguments.

Where the value being printed is an unsigned 64-bit value, use PRIu64
instead of %ld.

Add a cast when comparing an enum to an integer value.

Add an explicit uint64_t cast to the definition of BLOCK_CNT to prevent
value truncation.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/5137
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/unit/lib/bdev/vbdev_zone_block.c/vbdev_zone_block_ut.c
@@ -40,7 +40,7 @@
 #include "bdev/zone_block/vbdev_zone_block.c"
 #include "bdev/zone_block/vbdev_zone_block_rpc.c"
 
-#define BLOCK_CNT (1024ul * 1024ul * 1024ul * 1024ul)
+#define BLOCK_CNT ((uint64_t)1024ul * 1024ul * 1024ul * 1024ul)
 #define BLOCK_SIZE 4096
 
 /* Globals */
